protocol buffer type compatability changes:
I dwelled on this change for a while before going this route. Basically we cannot use uint8/uint16/IP address types when we serialize to a protocol buffer. Dict has some similar limitations where we parse the data type and switch before adding the value in. This change simplifies the data types throughout several important locations:
reporter.go -
	clientaddress/serveraddress - only convert these from net.IP to string for report logging
	time_stamp - convert this to an int type

stats.go -
	hops - just convert before sending to reports.LogEvent, the nfqueuemessage type uses a github.com/google/gopacket/layers dependency that pulls the type at that level

conntack.go -
	clientport/serverport/protocol/serverinterfaceid/serverinterfacetype/protocol - convert from uint8/uint16 to uint in the callbacks - Is it worth converting these to unsigned in the C side?

nfqueue.go -
	protocol/clientport/serverport/family - convert in the callback (see above for same C question)

classify.go -
	srcport/dstport - convert all the way at the tuple.go and session.go level and persist the types up from there, dict seems to handle the uint type just fine?
	productivity/risk - convert at the appclassmanager level, when we pull these out of
	the CSV we are already doing a uint8 to them, so just do uint instead

cloudmgr.go -
	protocol/serverport - convert this on the typing for fewer converts

dict.go -
	generateValue: add in a cases for int and uint, we are assuming these are all int32 when passed in... does this fly with our architectures? - Any int64s should be explicitly int64 in all other places

predicttrafficsvc.go -
	port/proto - convert these in any calls before we send to match up with tuple changes, and also on result parse output

+++ plugins/classify/cloudmgr.go
@@ -24,12 +24,12 @@ const cloudUploadInterval = 60 * time.Minute
 const cloudTableLimit = 1000
 
 type cloudReport struct {
-	Protocol    uint8  `json:"protocol"`
+	Protocol    uint   `json:"protocol"`
 	Application string `json:"application_control_application"`
 	Protochain  string `json:"application_control_protochain"`
 	Detail      string `json:"application_control_detail"`
 	ServerAddr  string `json:"s_server_addr"`
-	ServerPort  uint16 `json:"s_server_port"`
+	ServerPort  uint   `json:"s_server_port"`
 	Count       int    `json:"count"`
 }
